requirements.txt
praw
python-dotenv
huggingface_hub

